requirements.txt
beautifulsoup4
html5lib
pyaml
requests
